Add comprehensive logging utilities and tests for Starship Analyzer
- Implemented a logging configuration module to centralize logging setup.
- Created constants for log levels, formats, and file paths.
- Developed core logging functionalities including logger retrieval, global log level setting, and session management.
- Added custom colored formatter for console output.
- Implemented system information collection and logging, including CPU, memory, and GPU details.
- Created unit tests for logger utilities to ensure proper functionality and coverage.
- Ensured log directory creation and write access checks during initialization.
